Make a couple fixes for compilation errors/warnings.
- Change instances of `WOLFCLU_LOG(WOLFCLU_L0, "")` to `WOLFCLU_LOG(WOLFCLU_L0,
" ")` so that the compiler doesn't complain about 0-length format strings.
- In clu_funcs.c, I was getting warnings about the `XSTRNCPY` calls in
`wolfCLU_getAlgo`. It was warning that we weren't copying the null terminator.
I made the algo names const char arrays, and then we copy only up to the size
of the destination buffer. `strncpy` will 0-pad for us, so we don't need to
`memset` the destination buffer to 0s.