Improve x509 command.
Prior to this commit, x509 commands to convert between formats (or between the
same format, e.g. PEM to PEM) would potentially alter the data in the
certificate because of an intermediate transformation to a WOLFSSL_X509 object.
For example, converting a PEM cert to PEM would internally create a
WOLFSSL_X509 and then use that to generate the output. In some cases, the
extensions in the output cert would be in a different order than the input,
which is not ok because that invalidates the signature on the cert. This commit
makes it so any encoding conversion doesn't modify the cert.